+++ test/simple_animation_test.dart
@@ -33,4 +33,25 @@ void main() {
     expect(secondController.animationName, 'Animation 2');
     expect(secondController.mix, 0.8);
   });
+
+  test('SimpleAnimation exposes autoplay', () {
+    final firstController = SimpleAnimation(
+      riveFile.mainArtboard.animations.first.name,
+    );
+    // Autoplay defaults to true
+    expect(firstController.autoplay, isTrue);
+    // Controller should be active after being added to an artboard
+    riveFile.mainArtboard.addController(firstController);
+    expect(firstController.isActive, isTrue);
+
+    final secondController = SimpleAnimation(
+      riveFile.mainArtboard.animations.first.name,
+      autoplay: false,
+    );
+    // Autoplay defaults to true
+    expect(secondController.autoplay, isFalse);
+    // Controller should be inactive after being added to an artboard
+    riveFile.mainArtboard.addController(secondController);
+    expect(secondController.isActive, isFalse);
+  });
 }